requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
hypothesis>=6.82.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Caching (optional - can use in-memory for development)
redis>=4.6.0

# Environment management
python-dotenv>=1.0.0